moved `get_full_field` to the `laser_utils` as a stand-alone function of the laser object
