+++ lasy/laser.py
@@ -1,6 +1,5 @@
 import numpy as np
 import scipy.constants as scc
-from scipy.interpolate import interp1d
 from axiprop.lib import PropagatorFFT2, PropagatorResampling
 
 from lasy.utils.box import Box
@@ -272,78 +271,3 @@ class Laser:
             self.profile.lambda0,
             self.profile.pol,
         )
-
-    def get_full_field(self, theta=0, slice=0, slice_axis="x", Nt=None):
-        """
-        Reconstruct the laser pulse with carrier frequency on the default grid
-
-        Parameters
-        ----------
-        theta : float (rad) (optional)
-            Azimuthal angle
-        slice : float (optional)
-            Normalised position of the slice from -0.5 to 0.5
-        Nt: int (optional)
-            Number of time points on which field should be sampled. If is None,
-            the orignal time grid is used, otherwise field is interpolated on a
-            new grid.
-
-        Returns
-        -------
-            Et : ndarray (V/m)
-                The reconstructed field, with shape (Nr, Nt) (for `rt`)
-                or (Nx, Nt) (for `xyt`)
-            extent : ndarray (Tmin, Tmax, Xmin, Xmax)
-                Physical extent of the reconstructed field
-        """
-        omega0 = self.profile.omega0
-        field = self.field.field.copy()
-        time_axis = self.box.axes[-1]
-
-        if self.dim == "rt":
-            azimuthal_phase = np.exp(-1j * self.box.azimuthal_modes * theta)
-            field_upper = field * azimuthal_phase[:, None, None]
-            field_upper = field_upper.sum(0)
-            azimuthal_phase = np.exp(1j * self.box.azimuthal_modes * theta)
-            field_lower = field * azimuthal_phase[:, None, None]
-            field_lower = field_lower.sum(0)
-            field = np.vstack((field_lower[::-1][:-1], field_upper))
-        elif slice_axis == "x":
-            Nx_middle = field.shape[0] // 2 - 1
-            Nx_slice = int((1 + slice) * Nx_middle)
-            field = field[Nx_slice, :]
-        elif slice_axis == "y":
-            Ny_middle = field.shape[1] // 2 - 1
-            Ny_slice = int((1 + slice) * Ny_middle)
-            field = field[:, Ny_slice, :]
-        else:
-            return None
-
-        if Nt is not None:
-            Nr = field.shape[0]
-            time_axis_new = np.linspace(self.box.lo[-1], self.box.hi[-1], Nt)
-            field_new = np.zeros((Nr, Nt), dtype=field.dtype)
-
-            for ir in range(Nr):
-                interp_fu_abs = interp1d(time_axis, np.abs(field[ir]))
-                slice_abs = interp_fu_abs(time_axis_new)
-                interp_fu_angl = interp1d(time_axis, np.unwrap(np.angle(field[ir])))
-                slice_angl = interp_fu_angl(time_axis_new)
-                field_new[ir] = slice_abs * np.exp(1j * slice_angl)
-
-            time_axis = time_axis_new
-            field = field_new
-
-        field *= np.exp(-1j * omega0 * time_axis[None, :])
-        field = np.real(field)
-
-        if self.dim == "rt":
-            ext = np.array(
-                [self.box.lo[-1], self.box.hi[-1], -self.box.hi[0], self.box.hi[0]]
-            )
-        else:
-            ext = np.array(
-                [self.box.lo[-1], self.box.hi[-1], self.box.lo[0], self.box.hi[0]]
-            )
-
-        return field, ext
